requirements.txt
bcrypt==4.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
filelock==3.16.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
platformdirs==4.3.6
requests==2.32.3
urllib3==2.2.3
virtualenv==20.27.1
Werkzeug==3.1.3
zipp==3.21.0